Split router files
